fix: hide icon if description is missing
There might be future general improvements like merging the edit and info modal (since they both show the description, but only in one it's editable.

This PR already improves the situation a bit, since you don't have to click on that info button anymore to check __if__ there is a description at all.