fix(kanban): set new bucket id on task after moving it
This fixes a bug where a bucket update in the api was not triggered because the task still had the old bucket id saved. This led to inconsistencies between what was visible in the UI and what was actually stored.

Probably introduced in d68c62bd01580b578a9d8e994f64d889e8d22676

Resolves https://community.vikunja.io/t/warning-when-moving-tasks-to-a-bucket-with-limit/3520